# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Study Group Finder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: List all study groups
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [online, offline]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupWithMemberCount"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new study group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupBody"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Validation error or duplicate group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Daily group creation limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a study group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          description: Group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete a group (admin or creator only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a study group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Already a member or group full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{id}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a study group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /groups/{id}/members:
    get:
      operationId: getGroupMembers
      tags: [groups]
      summary: Get members of a study group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /groups/{id}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Message flagged for banned content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        faculty:
          type: string
        academicYear:
          type: integer
        subjectsOfInterest:
          type: array
          items:
            type: string
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - faculty
        - academicYear
        - isBanned
...
    Group:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [online, offline]
        location:
          type: string
          nullable: true
        dateTime:
          type: string
          format: date-time
        maxMembers:
          type: integer
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - subject
        - description
        - type
        - dateTime
        - maxMembers
        - createdBy
    GroupWithMemberCount:
      allOf:
        - $ref: "#/components/schemas/Group"
        - type: object
          properties:
            memberCount:
              type: integer
            creator:
              $ref: "#/components/schemas/User"
            isMember:
              type: boolean
          required:
            - memberCount
    GroupDetail:
      allOf:
        - $ref: "#/components/schemas/GroupWithMemberCount"
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/User"
          required:
            - members
    CreateGroupBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [online, offline]
        location:
          type: string
          nullable: true
        dateTime:
          type: string
          format: date-time
        maxMembers:
          type: integer
      required:
        - title
        - subject
        - description
        - type
        - dateTime
        - maxMembers
    Message:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        userId:
          type: integer
        content:
          type: string
        isFlagged:
          type: boolean
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - groupId
        - userId
        - content
        - isFlagged
        - createdAt
...
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...